refactor: Create file to store Color model
